--- shopify/shop-mini-globe-demo/CLAUDE.md
@@ -1,1 +1,0 @@
-AGENTS.md
