+++ shopify/shop-mini-globe-demo/minis/shop-mini-globe-demo/src/types.ts
@@ -1,0 +1,3 @@
+export type PathPoint = { location: string; lat: number; lng: number };
+export type HistoryShipment = { id: string; trackingNumber: string; path: PathPoint[] };
+export type HistoryStats = { packages: number; cities: number; km: number };
